Split Application and API into two distinct modules
